Refactor Windows Store domain knowledge out of the locator
