+++ src/client/pythonEnvironments/base/locators/lowLevel/windowsKnownPathsLocator.ts
@@ -9,7 +9,7 @@ import { getSearchPathEntries } from '../../../../common/utils/exec';
 import { Disposables, IDisposable } from '../../../../common/utils/resourceLifecycle';
 import { iterPythonExecutablesInDir, looksLikeBasicGlobalPython } from '../../../common/commonUtils';
 import { isPyenvShimDir } from '../../../common/environmentManagers/pyenv';
-import { isWindowsStoreDir } from './windowsStoreLocator';
+import { isWindowsStoreDir } from '../../../common/environmentManagers/windowsStoreEnv';
 import { PythonEnvKind, PythonEnvSource } from '../../info';
 import { BasicEnvInfo, ILocator, IPythonEnvsIterator, PythonLocatorQuery } from '../../locator';
 import { Locators } from '../../locators';
