+++ src/client/pythonEnvironments/common/commonUtils.ts
@@ -10,7 +10,7 @@ import { PythonVersion, UNKNOWN_PYTHON_VERSION } from '../base/info';
 import { comparePythonVersionSpecificity } from '../base/info/env';
 import { parseVersion } from '../base/info/pythonVersion';
 import { getPythonVersionFromConda } from './environmentManagers/conda';
-import { getPythonVersionFromPyvenvCfg } from './environmentManagers/simplevirtualenvs';
+import { getPythonVersionFromPyvenvCfg } from './environmentManagers/simpleVirtualEnvs';
 import * as posix from './posixUtils';
 import * as windows from './windowsUtils';
